+++ src/main/java/ch/martinelli/demo/hilla/repository/PersonRepository.java
@@ -4,6 +4,6 @@ import ch.martinelli.demo.hilla.entity.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface PersonRepository extends JpaRepository<Person, Long>, JpaSpecificationExecutor<Person> {
+public interface PersonRepository extends JpaRepository<Person, Integer>, JpaSpecificationExecutor<Person> {
 
 }
